backend/requirements.txt
requests
pandas
python-dotenv
spacy
geopy
tqdm
fastapi
uvicorn
google-generativeai
